requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# OCR Engines
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.3
paddlepaddle==2.5.2

# PDF Processing
PyPDF2==3.0.1
reportlab==4.0.7
pdf2image==1.16.3
pypdfium2==4.26.0

# NLP and AI
spacy==3.7.2
transformers==4.35.2
torch==2.1.1

# Utilities
python-dateutil==2.8.2
regex==2023.10.3
pydantic==2.5.2

# Progress and Logging
tqdm==4.66.1
loguru==0.7.2
